functions/azure/requirements.txt
azure-functions>=1.12.0
azure-identity>=1.10.0
azure-mgmt-authorization>=2.0.0
azure-mgmt-resource>=21.1.0
azure-core>=1.24.0
python-dotenv==0.21.1
cryptography==38.0.4
cffi==1.15.1
pycparser==2.21
requests>=2.28.0
msgraph-core>=0.2.2
msrest>=0.7.1
msgraph-sdk==1.0.0
